[Fixes #6117] Added RedirectToPage overloads to Controller
