+++ src/Microsoft.AspNetCore.Mvc.Core/DependencyInjection/MvcCoreServiceCollectionExtensions.cs
@@ -215,6 +215,7 @@ namespace Microsoft.Extensions.DependencyInjection
             services.TryAddSingleton<LocalRedirectResultExecutor>();
             services.TryAddSingleton<RedirectToActionResultExecutor>();
             services.TryAddSingleton<RedirectToRouteResultExecutor>();
+            services.TryAddSingleton<RedirectToPageResultExecutor>();
             services.TryAddSingleton<ContentResultExecutor>();
 
             //
