+++ src/Microsoft.AspNetCore.Mvc.RazorPages/DependencyInjection/MvcRazorPagesMvcCoreBuilderExtensions.cs
@@ -102,7 +102,6 @@ namespace Microsoft.Extensions.DependencyInjection
 
             // Action executors
             services.TryAddSingleton<PageResultExecutor>();
-            services.TryAddSingleton<RedirectToPageResultExecutor>();
 
             services.TryAddTransient<PageSaveTempDataPropertyFilter>();
         }
