+++ src/Microsoft.AspNetCore.Mvc.RazorPages/Resources.resx
@@ -141,7 +141,4 @@
   <data name="PathMustBeAnAppRelativePath" xml:space="preserve">
     <value>Path must be an application relative path that starts with a forward slash '/'.</value>
   </data>
-  <data name="NoRoutesMatched" xml:space="preserve">
-    <value>No page named '{0}' matches the supplied values.</value>
-  </data>
 </root>
