+++ test/WebSites/RazorPagesWebSite/RedirectToController.cshtml
@@ -3,6 +3,6 @@
 @functions {
     public IActionResult OnGet()
     {
-        return new RedirectToRouteResult(new { controller = "Redirect", action = "RedirectToPage", param = 92 });
+        return new RedirectToRouteResult(new { controller = "Redirect", action = "RedirectToPageAction", param = 92 });
     }
 }
